Clean up deprecated `gulp` task runner
