+++ build-system/task-runner/amp-task-runner.js
@@ -28,9 +28,9 @@ const {
   updatePackages,
   updateSubpackages,
 } = require('../common/update-packages');
-const {cyan, red, green, magenta, yellow} = require('kleur/colors');
+const {cyan, red, green, magenta} = require('kleur/colors');
 const {isCiBuild} = require('../common/ci');
-const {log, logWithoutTimestamp} = require('../common/logging');
+const {log} = require('../common/logging');
 
 /**
  * Special-case constant that indicates if `amp --help` was invoked.
@@ -188,37 +188,7 @@ function finalizeRunner() {
   commander.parse();
 }
 
-/**
- * Prints a deprecation notice for the gulp task runner.
- * @param {boolean} withTimestamps
- */
-function printGulpDeprecationNotice(withTimestamps) {
-  const logFunc = withTimestamps ? log : logWithoutTimestamp;
-  logFunc(yellow('=*='.repeat(25)));
-  logFunc(yellow('DEPRECATION NOTICE:'));
-  logFunc(
-    'All',
-    cyan('gulp'),
-    'tasks have been replaced by an identical set of',
-    cyan('amp'),
-    'tasks.'
-  );
-  logFunc('⤷ Run', cyan('amp --help'), 'for a full list of tasks.');
-  logFunc(
-    '⤷ Run',
-    cyan('amp <command> --help'),
-    'for help with a specific task.'
-  );
-  logFunc(
-    '⤷ See',
-    cyan('contributing/TESTING.md#testing-commands'),
-    'for more info.'
-  );
-  logFunc(yellow('=*='.repeat(25)));
-}
-
 module.exports = {
   createTask,
   finalizeRunner,
-  printGulpDeprecationNotice,
 };
