--- gulp-deprecated.js
@@ -1,27 +1,0 @@
-#!/usr/bin/env node
-/**
- * Copyright 2021 The AMP HTML Authors. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS-IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-const {
-  printGulpDeprecationNotice,
-} = require('./build-system/task-runner/amp-task-runner');
-
-/**
- * Print a deprecation notice and fall back to the amp task runner.
- * TODO(amphtml): Remove this script after a month or so.
- */
-printGulpDeprecationNotice(/* withTimeStamps */ true);
-require('./amp');
