+++ package.json
@@ -12,7 +12,7 @@
   },
   "scripts": {
     "preinstall": "node build-system/common/check-package-manager.js",
-    "postinstall": "node build-system/task-runner/install-task-runner.js"
+    "postinstall": "node build-system/task-runner/install-amp-task-runner.js"
   },
   "dependencies": {
     "@ampproject/animations": "0.2.2",
